@@ -1,8 +1,9 @@
 <?php
 
-namespace Behat\Mink\Tests\Driver;
+namespace SilverStripe\MinkSelenium3Driver\Tests;
 
-use Behat\Mink\Driver\Selenium2Driver;
+use Behat\Mink\Tests\Driver\AbstractConfig;
+use SilverStripe\MinkSelenium3Driver\Selenium3Driver;
 
 class Selenium2Config extends AbstractConfig
 {
@@ -19,7 +20,7 @@ class Selenium2Config extends AbstractConfig
         $browser = getenv('WEB_FIXTURES_BROWSER') ?: 'firefox';
         $seleniumHost = $_SERVER['DRIVER_URL'];
 
-        return new Selenium2Driver($browser, null, $seleniumHost);
+        return new Selenium3Driver($browser, null, $seleniumHost);
     }
 
     /**
